imported 20060531 current

+++ src/sbin/fsck_lfs/pass4.c
@@ -1,4 +1,4 @@
-/* $NetBSD: pass4.c,v 1.11 2005/02/06 06:13:47 perry Exp $	 */
+/* $NetBSD: pass4.c,v 1.13 2005/09/13 04:14:17 christos Exp $	 */
 
 /*
  * Copyright (c) 1980, 1986, 1993
@@ -45,7 +45,7 @@
 
 #include "bufcache.h"
 #include "vnode.h"
-#include "lfs.h"
+#include "lfs_user.h"
 
 #include "fsutil.h"
 #include "fsck.h"
@@ -106,8 +106,8 @@ pass4(void)
 			break;
 
 		default:
-			err(8, "BAD STATE %d FOR INODE I=%d\n",
-			    statemap[inumber], inumber);
+			err(8, "BAD STATE %d FOR INODE I=%llu\n",
+			    statemap[inumber], (unsigned long long)inumber);
 		}
 	}
 }
